allowing Python sys.argv to obtain arguments from dlvhex commandline
An alternative would be to pass all filenames from dlvhex to Python, or
all filenames and all arguments after a certain argument.

However getopt() can reorder arguments so the meaningfulness of "after
an argument" is not clear.

The current solution allows to use --pythonarg=<arg> one or more times,
and all these arguments end up in sys.argv. The first element of
sys.argv is the python script name.